Move Collection Form into Component
