+++ app/src/app/(app)/(screens)/(collection)/create.tsx
@@ -1,4 +1,3 @@
-import { useForm, Controller } from "react-hook-form";
 import { useSession } from "@/src/context/AuthContext";
 import { useMutation } from "@tanstack/react-query";
 import {
@@ -8,37 +7,10 @@ import {
 import { toISODateString } from "@/src/libs/date";
 import { router } from "expo-router";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { View } from "react-native";
-import {
-  FormControl,
-  FormControlError,
-  FormControlErrorIcon,
-  FormControlErrorText,
-  FormControlLabel,
-  FormControlLabelText,
-} from "@/src/components/ui/form-control";
-import { Heading } from "@/src/components/ui/heading";
-import { Input, InputField } from "@/src/components/ui/input";
-import { Textarea, TextareaInput } from "@/src/components/ui/textarea";
-import { Button, ButtonSpinner, ButtonText } from "@/src/components/ui/button";
-import { useCallback, useState } from "react";
-import { AlertCircleIcon } from "@/src/components/ui/icon";
-import LocationInput, {
-  GeoLocation,
-} from "@/src/components/inputs/LocationInput";
-import { FlatList } from "react-native";
 import { queryClient } from "@/src/app/_layout";
-import DatePickerInput from "@/src/components/inputs/DatePickerInput";
-
-/**
- * Form values for the CreateCollection screen
- */
-interface CreateCollectionForm {
-  title: string;
-  date: Date | null;
-  location: GeoLocation;
-  caption: string;
-}
+import CollectionForm, {
+  CollectionFormData,
+} from "@/src/components/forms/CollectionForm";
 
 /**
  * @description Screen for creating a new collection
@@ -50,21 +22,6 @@ interface CreateCollectionForm {
  */
 export default function CreateCollection() {
   const { session } = useSession();
-  const {
-    control,
-    handleSubmit,
-    watch,
-    setValue,
-    formState: { errors },
-  } = useForm<CreateCollectionForm>({
-    defaultValues: {
-      title: "",
-      caption: "",
-      date: null,
-      location: { text: "" },
-    },
-  });
-
   const createMutation = useMutation(
     createNewCollectionApiUserUserIdCollectionPostMutation(),
   );
@@ -75,7 +32,7 @@ export default function CreateCollection() {
    *
    * @param {CreateCollectionForm} form - Form containing collection detals
    */
-  const onSubmit = async (form: CreateCollectionForm) => {
+  const onSubmit = async (form: CollectionFormData) => {
     const {
       location: { lat, long, text },
       date,
@@ -112,134 +69,13 @@ export default function CreateCollection() {
     );
   };
 
-  // Prevent re-rendering location input when Geolocation changes
-  const locationValue = watch("location");
-
-  /**
-   * Updates the location input when GeoLocation changes
-   *
-   * @param {GeoLocation} location - new location value
-   */
-  const handleLocationChange = useCallback(
-    (location: GeoLocation) => {
-      const hasChanged =
-        locationValue.text !== location.text ||
-        locationValue.lat !== location.lat ||
-        locationValue.long !== location.long;
-
-      if (hasChanged) {
-        setValue("location", location);
-      }
-    },
-    [locationValue, setValue],
-  );
-
   return (
     <SafeAreaView className="flex-1" edges={["bottom"]}>
-      <FlatList
-        data={[]}
-        renderItem={() => <></>}
-        contentContainerStyle={{ paddingHorizontal: 20, paddingTop: 20 }}
-        showsVerticalScrollIndicator={false}
-        keyboardShouldPersistTaps="handled"
-        ListHeaderComponent={
-          <View className="flex justify-center gap-6">
-            <Heading className="block" size="2xl">
-              Create Collection
-            </Heading>
-            <FormControl size={"lg"} isInvalid={!!errors?.title}>
-              <FormControlLabel>
-                <FormControlLabelText>Title</FormControlLabelText>
-              </FormControlLabel>
-              <Controller
-                name="title"
-                control={control}
-                render={({ field }) => (
-                  <Input className="bg-background-0">
-                    <InputField
-                      onChangeText={field.onChange}
-                      value={field.value}
-                      placeholder="Collection Title"
-                    />
-                  </Input>
-                )}
-                rules={{
-                  validate: {
-                    required: (value) => {
-                      return (value && value.length > 0) || "Title is required";
-                    },
-                  },
-                }}
-              />
-              <FormControlError className="mt-4">
-                <FormControlErrorIcon as={AlertCircleIcon} />
-                <FormControlErrorText className="flex-1">
-                  {errors?.title?.message}
-                </FormControlErrorText>
-              </FormControlError>
-            </FormControl>
-            <FormControl size={"lg"}>
-              <FormControlLabel>
-                <FormControlLabelText>Caption</FormControlLabelText>
-              </FormControlLabel>
-              <Controller
-                name="caption"
-                control={control}
-                render={({ field }) => (
-                  <Textarea className="bg-background-0" size="md">
-                    <TextareaInput
-                      onChangeText={field.onChange}
-                      value={field.value ?? ""}
-                      placeholder="Add a caption"
-                    />
-                  </Textarea>
-                )}
-              />
-            </FormControl>
-            <FormControl size={"lg"}>
-              <FormControlLabel>
-                <FormControlLabelText>Date</FormControlLabelText>
-              </FormControlLabel>
-              <Controller
-                name="date"
-                control={control}
-                render={({ field }) => (
-                  <DatePickerInput
-                    value={field.value}
-                    onChange={(date) => field.onChange(date)}
-                  />
-                )}
-              />
-            </FormControl>
-            <FormControl size={"lg"}>
-              <FormControlLabel>
-                <FormControlLabelText>Location</FormControlLabelText>
-              </FormControlLabel>
-              <Controller
-                name="location"
-                control={control}
-                render={({ field }) => (
-                  <LocationInput
-                    onChange={handleLocationChange}
-                    value={field.value}
-                  />
-                )}
-              />
-            </FormControl>
-            <Button
-              className="mt-auto"
-              size={"lg"}
-              onPress={handleSubmit(onSubmit)}
-              disabled={createMutation.isPending}
-            >
-              {createMutation.isPending ? (
-                <ButtonSpinner />
-              ) : (
-                <ButtonText>Create Collection</ButtonText>
-              )}
-            </Button>
-          </View>
-        }
+      <CollectionForm
+        title="Create Collection"
+        submitButtonText="Create Collection"
+        isSubmitting={createMutation.isPending}
+        onSubmit={onSubmit}
       />
     </SafeAreaView>
   );
